# Enhance/perf optimizations