--- a/Lib/gc_tweak.py
+++ b/Lib/gc_tweak.py
@@ -4,12 +4,32 @@
 # Based on article
 # https://mkennedy.codes/posts/python-gc-settings-change-this-and-make-your-app-go-20pc-faster/
 import gc
+import time
 import psutil
 
 # Get the current process to measure CPU usage
 process = psutil.Process()
+total_gc_time = 0
+gc_start_time = None
+app_start_time = time.perf_counter()  # Track when the app starts
 
-def optimize_gc(allocs: int = 50_000, gen1_factor: int = 2, gen2_factor: int = 2, freeze: bool = True):
+def track_gc_time(phase, info):
+    global total_gc_time, gc_start_time
+
+    if phase == "start":
+        gc_start_time = time.perf_counter()
+    elif phase == "stop" and gc_start_time is not None:
+        total_gc_time += time.perf_counter() - gc_start_time
+        gc_start_time = None  # Reset after use
+
+        # Calculate GC overhead percentage
+        elapsed_time = time.perf_counter() - app_start_time
+        gc_percentage = (total_gc_time / elapsed_time) * 100 if elapsed_time > 0 else 0
+
+        print(f"Total GC time: {total_gc_time:.6f}s ({gc_percentage:.2f}% of runtime)")
+
+def optimize_gc(allocs: int = 50_000, gen1_factor: int = 2, gen2_factor: int = 2,
+                freeze: bool = True, show_data = False):
     """
     Optimize Python's garbage collection settings for performance.
 
@@ -33,4 +53,7 @@ def optimize_gc(allocs: int = 50_000, gen1_factor: int = 2, gen2_factor: int = 2
     new_gen1 = gen1 * gen1_factor
     new_gen2 = gen2 * gen2_factor
     gc.set_threshold(allocs, new_gen1, new_gen2)
+
+    if show_data:
+        gc.callbacks.append(track_gc_time)