--- a/Scripts/custom_status_bar.py
+++ b/Scripts/custom_status_bar.py
@@ -20,9 +20,11 @@ from enum import Enum
 from dataclasses import dataclass, field
 from typing import Any, ClassVar, Optional
 
+import csv
 import psutil
 import requests
 from SimConnect import SimConnect, AircraftRequests
+import subprocess
 
 try:
     # Import all color print functions
@@ -952,23 +954,40 @@ def convert_real_world_time_to_sim_time(real_world_time):
 
 # --- SimConnect Lookup Functions ----------------------------------------------------------------
 def is_sim_running(min_runtime=120):
-    """Check if Microsoft Flight Simulator is running"""
-    now = time.time()  # Get current time
+    """Return True if an MSFS process has been running for at least min_runtime seconds."""
+    try:
+        cmd = (
+            'wmic process where "name like \'FlightSimulator%%.exe\'" '
+            'get Name,CreationDate,ProcessId /format:csv'
+        )
+        output = subprocess.check_output(cmd, shell=True).decode(errors="ignore").strip()
+    except subprocess.CalledProcessError:
+        return False
 
-    for process in psutil.process_iter(['name', 'create_time']):
-        process_name = process.info['name']
-        start_time = process.info.get('create_time', 0)
+    now = time.time()
+    reader = csv.DictReader(StringIO(output))
 
-        if process_name and process_name.lower().startswith("flightsimulator") \
-          and process_name.lower().endswith(".exe"):
-            runtime = now - start_time  # Calculate how long the process has been running
+    for row in reader:
+        name = row.get("Name", "").strip()
+        creation = row.get("CreationDate", "").split('.')[0].strip()
+        pid = row.get("ProcessId", "").strip()
 
-            if runtime >= min_runtime:
-                print_debug(f"Found MSFS process: {process_name} (Running for {runtime:.1f} sec)")
-                return True
-            else:
-                print_debug(f"Found MSFS process: {process_name}, "
-                            f"but only running for {runtime:.1f} sec (Waiting...)")
+        if not (name.startswith("FlightSimulator") and creation and pid.isdigit()):
+            continue
+
+        try:
+            start_time = time.mktime(time.strptime(creation, "%Y%m%d%H%M%S"))
+        except ValueError:
+            print(f"[ERROR] Could not parse creation time: {creation}")
+            continue
+
+        runtime = now - start_time
+        if runtime >= min_runtime:
+            print(f"Found MSFS process: {name} (PID: {pid}, Running for {runtime:.1f} sec)")
+            return True
+        else:
+            print(f"Found {name} (PID: {pid}), but only running for {runtime:.1f} sec (Waiting...)")
+            return False
 
     return False