## Performance Optimizations

This branch has been focused on improving performance on some of the scripts, particularly CPU use.  It has also added some tools to help with this as well.

### Benchmarking Methodology

- CPU usage is measured as the **average total CPU usage over a 2.5-minute period**.
- The following table compares CPU usage **before and after** optimizations.

### Before vs. After Optimization Results

| Script                  | CPU Usage (Before) | CPU Usage (After) | Relative Reduction |
|-------------------------|-------------------|-------------------|---------------------|
| fenix_radio.py          | 0.32%             | 0.01%             | 🔽 96.88%          |
| custom_status_bar.py    | 0.26%             | 0.12%             | 🔽 53.85%          |
| plot_joystick.py        | 0.31%             | 0.16%             | 🔽 48.39%          |

## Launcher: Performance Monitor Improvements
- Improved **CPU usage measurement accuracy**.
- Now uses a **short-term moving average**, which better reflects real-time CPU usage, aligning more closely with Windows Task Manager.

## Other Changes
- Added gc_tweak.py lib: A utility to allow easy tweaks of Python garbage collector.
- Added tk_profiler.py lib: A utility to allow a profiler to be attached to Tkinter root window to profile measured spike in CPU activity.